Name Powershell Script and Log Files Uniquely
In order to allow winrm-elevated to run concurrently against the
same Windows server, name the powershell script uploaded to the server
using a randomly generated UUID.

In addition, the logfiles generated by the script will be named using the
GetTempFileName to ensure uniqueness.

We will clean up the logfiles when the script is finished. Since the same
script can be uploaded once and invoked multiple times it will be left intact.