Port the hash case for the erubis->erubi change from WinRM
Adapted the change from:
https://github.com/WinRb/WinRM/pull/300

This grabs prior context or local variables, builds a string which set
these variables, and evalulates them in the context of the binding.

Note, this repo wasn't explicitly depending on erubis before but was trying to
require it.  I've also added erubi as a dependency so this gem can run
standalone.

Fixes #28